+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Publish artifacts for the `JS` and `wasmJs` targets for the jetpack-compose extensions.
 - Added support to `formatCssString` and `Color.parse` for color spaces added in recent updates to the CSS color spec: `oklab`, `oklch`, `srgb-linear`, `xyz-d50` and `xyz-d65`. 
+- Added `customColorSpaces` for `Color.parse` and `formatCssString` to allow non-standard color spaces to be used in color strings.
 
 ### Changed
 - `Color.parse` now parses `lch()` and `lab()` functions with the with D50 white points instead of D65 in order to comply with the CSS color spec.